Chapter05/translate-api/requirements.txt
google-cloud-translate==1.3.0
tweepy==3.6.0
